refactor(frontend): Split component files in separate dirs
